+++ components/Navbar/index.tsx
@@ -4,6 +4,11 @@ import styles from '../../styles/Navbar.module.css'
 export default function Navbar() {
     return (
         <div className={styles.navbar}>
+            <Link href='/'>
+                <a className={styles.navlink}>
+                    Dashboard
+                </a>
+            </Link>
             <Link href='/current-readings'>
                 <a className={styles.navlink}>
                     Current Readings