+++ pages/index.tsx
@@ -1,10 +1,14 @@
 import type { NextPage } from 'next'
+import type { GetStaticProps } from 'next'
 import Layout from '../components/layout'
 import styles from '../styles/Home.module.css'
 
 import Navbar from '../components/Navbar'
 
 const Home: NextPage = () => {
+
+  const apiData = 'username';
+
   return (
     <Layout>
       <aside className={styles.aside}>
@@ -12,10 +16,19 @@ const Home: NextPage = () => {
       </aside>
 
       <section className={styles.home}>
-        This is the home page
+        Welcome, {apiData}
       </section>
     </Layout>
   )
 }
+
+// export const getStaticProps: GetStaticProps = async () => {
+//   const apiData = 'username';
+//   return {
+//     props: {
+//       apiData
+//     }
+//   }
+// }
 
 export default Home